requirement.txt
streamlit
openai
langchain
langchain-community
faiss-cpu
youtube_transcript_api
python-dotenv
tiktoken